Make Windows 7/8/10 download intuitively accesable
A friend and I have noticed (or rather, not noticed) the Windows 7/8/10 downloads intuitively, since the title is the first thing and the version name is bold, we just assumed its part of a title (like "Hexchat Downloads"). Now that it can be found and the version name is no longer bold, I hope it is intuitively findable.